LayoutTests/http/tests/xmlhttprequest/auth-reject-protection-space-expected.txt

Simulating reject protection space and continue for authentication challenge
ALERT: status: 401
